+++ server/core/src/https/extractors/mod.rs
@@ -2,7 +2,7 @@ use axum::{
     async_trait,
     extract::{
         connect_info::{ConnectInfo, Connected},
-        FromRequest, FromRequestParts,
+        FromRequestParts,
     },
     http::{
         header::{HeaderName, AUTHORIZATION as AUTHORISATION},
@@ -176,99 +176,84 @@ impl FromRequestParts<ServerState> for VerifiedClientInformation {
     }
 }
 
-#[async_trait]
-impl FromRequest<ServerState, axum::body::Body>
-    for axum::Form<kanidm_proto::oauth2::AccessTokenRequest>
-{
-    type Rejection = &'static str;
+// #[async_trait]
+// impl FromRequest<ServerState, axum::body::Body> for VerifiedClientInformation {
+//     type Rejection = &'static str;
 
-    // #[instrument(level = "debug", skip(state))]
-    async fn from_request(
-        _request: axum::extract::Request,
-        _state: &ServerState,
-    ) -> Result<Self, Self::Rejection> {
-        todo!()
-    }
-}
+//     #[instrument(level = "debug", skip(state))]
+//     async fn from_request(
+//         request: axum::extract::Request,
+//         state: &ServerState,
+//     ) -> Result<Self, Self::Rejection> {
+//         let (parts, _body) = request.into_parts();
+//         let ConnectInfo(ClientConnInfo { addr, client_cert }) = parts
+//             .extract::<ConnectInfo<ClientConnInfo>>()
+//             .await
+//             .map_err(|_| {
+//                 error!("Connect info contains invalid data");
+//                 // (
+//                 // StatusCode::BAD_REQUEST,
+//                 "connect info contains invalid data"
+//                 // ,
+//                 // )
+//             })?;
 
-#[async_trait]
-impl FromRequest<ServerState, axum::body::Body> for VerifiedClientInformation {
-    type Rejection = &'static str;
+//         let ip_addr = if state.trust_x_forward_for {
+//             if let Some(x_forward_for) = parts.headers.get(X_FORWARDED_FOR_HEADER) {
+//                 // X forward for may be comma separated.
+//                 let first = x_forward_for
+//                     .to_str()
+//                     .map(|s|
+//                         // Split on an optional comma, return the first result.
+//                         s.split(',').next().unwrap_or(s))
+//                     .map_err(|_| {
+//                         // (
+//                         // StatusCode::BAD_REQUEST,
+//                         "X-Forwarded-For contains invalid data"
+//                         // ,
+//                         // )
+//                     })?;
 
-    #[instrument(level = "debug", skip(state))]
-    async fn from_request(
-        request: axum::extract::Request,
-        state: &ServerState,
-    ) -> Result<Self, Self::Rejection> {
-        let (parts, _body) = request.into_parts();
-        let ConnectInfo(ClientConnInfo { addr, client_cert }) = parts
-            .extract::<ConnectInfo<ClientConnInfo>>()
-            .await
-            .map_err(|_| {
-                error!("Connect info contains invalid data");
-                // (
-                // StatusCode::BAD_REQUEST,
-                "connect info contains invalid data"
-                // ,
-                // )
-            })?;
+//                 first.parse::<IpAddr>().map_err(|_| {
+//                     // (
+//                     // StatusCode::BAD_REQUEST,
+//                     "X-Forwarded-For contains invalid ip addr"
+//                     // ,
+//                     // )
+//                 })?
+//             } else {
+//                 addr.ip()
+//             }
+//         } else {
+//             addr.ip()
+//         };
 
-        let ip_addr = if state.trust_x_forward_for {
-            if let Some(x_forward_for) = parts.headers.get(X_FORWARDED_FOR_HEADER) {
-                // X forward for may be comma separated.
-                let first = x_forward_for
-                    .to_str()
-                    .map(|s|
-                        // Split on an optional comma, return the first result.
-                        s.split(',').next().unwrap_or(s))
-                    .map_err(|_| {
-                        // (
-                        // StatusCode::BAD_REQUEST,
-                        "X-Forwarded-For contains invalid data"
-                        // ,
-                        // )
-                    })?;
+//         let bearer_token = if let Some(header) = parts.headers.get(AUTHORISATION) {
+//             header
+//                 .to_str()
+//                 .map_err(|err| {
+//                     warn!(?err, "Invalid bearer token, ignoring");
+//                 })
+//                 .ok()
+//                 .and_then(|s| s.split_once(' '))
+//                 .map(|(_, s)| s.to_string())
+//                 .or_else(|| {
+//                     warn!("bearer token format invalid, ignoring");
+//                     None
+//                 })
+//         } else {
+//             None
+//         };
 
-                first.parse::<IpAddr>().map_err(|_| {
-                    // (
-                    // StatusCode::BAD_REQUEST,
-                    "X-Forwarded-For contains invalid ip addr"
-                    // ,
-                    // )
-                })?
-            } else {
-                addr.ip()
-            }
-        } else {
-            addr.ip()
-        };
-
-        let bearer_token = if let Some(header) = parts.headers.get(AUTHORISATION) {
-            header
-                .to_str()
-                .map_err(|err| {
-                    warn!(?err, "Invalid bearer token, ignoring");
-                })
-                .ok()
-                .and_then(|s| s.split_once(' '))
-                .map(|(_, s)| s.to_string())
-                .or_else(|| {
-                    warn!("bearer token format invalid, ignoring");
-                    None
-                })
-        } else {
-            None
-        };
-
-        Ok(VerifiedClientInformation(ClientAuthInfo {
-            source: Source::Https(ip_addr),
-            client_cert,
-            bearer_token: None,
-            // bearer_token, // TODO
-            basic_authz: None,
-        }))
-    }
-}
+//         Ok(VerifiedClientInformation(ClientAuthInfo {
+//             source: Source::Https(ip_addr),
+//             client_cert,
+//             bearer_token: None,
+//             // bearer_token, // TODO
+//             basic_authz: None,
+//         }))
+//     }
+// }
 
 #[derive(Debug, Clone)]
 pub struct ClientConnInfo {
